move DokkatooFormatPlugin util classes into separate files
