+++ modules/dokkatoo-plugin/src/main/kotlin/formats/DokkatooFormatPlugin.kt
@@ -5,33 +5,12 @@ import dev.adamko.dokkatoo.DokkatooExtension
 import dev.adamko.dokkatoo.adapters.DokkatooAndroidAdapter
 import dev.adamko.dokkatoo.adapters.DokkatooJavaAdapter
 import dev.adamko.dokkatoo.adapters.DokkatooKotlinAdapter
-import dev.adamko.dokkatoo.distributions.DokkatooConfigurationAttributes
-import dev.adamko.dokkatoo.distributions.DokkatooConfigurationAttributes.Companion.DOKKATOO_BASE_ATTRIBUTE
-import dev.adamko.dokkatoo.distributions.DokkatooConfigurationAttributes.Companion.DOKKATOO_CATEGORY_ATTRIBUTE
-import dev.adamko.dokkatoo.distributions.DokkatooConfigurationAttributes.Companion.DOKKA_FORMAT_ATTRIBUTE
-import dev.adamko.dokkatoo.dokka.DokkaPublication
 import dev.adamko.dokkatoo.internal.*
-import dev.adamko.dokkatoo.tasks.DokkatooGenerateTask
-import dev.adamko.dokkatoo.tasks.DokkatooPrepareModuleDescriptorTask
-import dev.adamko.dokkatoo.tasks.DokkatooPrepareParametersTask
 import javax.inject.Inject
-import org.gradle.api.NamedDomainObjectProvider
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.Dependency
 import org.gradle.api.artifacts.dsl.DependencyHandler
-import org.gradle.api.attributes.AttributeContainer
-import org.gradle.api.attributes.Bundling.BUNDLING_ATTRIBUTE
-import org.gradle.api.attributes.Bundling.EXTERNAL
-import org.gradle.api.attributes.Category.CATEGORY_ATTRIBUTE
-import org.gradle.api.attributes.Category.LIBRARY
-import org.gradle.api.attributes.LibraryElements.JAR
-import org.gradle.api.attributes.LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE
-import org.gradle.api.attributes.Usage.JAVA_RUNTIME
-import org.gradle.api.attributes.Usage.USAGE_ATTRIBUTE
-import org.gradle.api.attributes.java.TargetJvmEnvironment.STANDARD_JVM
-import org.gradle.api.attributes.java.TargetJvmEnvironment.TARGET_JVM_ENVIRONMENT_ATTRIBUTE
 import org.gradle.api.file.FileSystemOperations
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
@@ -79,13 +58,13 @@ abstract class DokkatooFormatPlugin(
       val dokkatooConsumer =
         target.configurations.named(DokkatooBasePlugin.dependencyContainerNames.dokkatoo)
 
-      val dependencyContainers = DependencyContainers(
+      val dependencyContainers = DokkatooFormatDependencyContainers(
         formatName = formatName,
         dokkatooConsumer = dokkatooConsumer,
         project = target,
       )
 
-      val dokkatooTasks = DokkatooTasks(
+      val dokkatooTasks = DokkatooFormatTasks(
         project = target,
         publication = publication,
         dokkatooExtension = dokkatooExtension,
@@ -137,7 +116,7 @@ abstract class DokkatooFormatPlugin(
   class DokkatooFormatPluginContext(
     val project: Project,
     val dokkatooExtension: DokkatooExtension,
-    val dokkatooTasks: DokkatooTasks,
+    val dokkatooTasks: DokkatooFormatTasks,
     formatName: String,
   ) {
     private val dependencyContainerNames = DokkatooBasePlugin.DependencyContainerNames(formatName)
@@ -194,262 +173,4 @@ abstract class DokkatooFormatPlugin(
     }
   }
 
-
-  /**
-   * The Dokka-specific Gradle [Configuration]s used to produce and consume files from external sources
-   * (example: Maven Central), or between subprojects.
-   *
-   * (Be careful of the confusing names: Gradle [Configuration]s are used to transfer files,
-   * [DokkaConfiguration][org.jetbrains.dokka.DokkaConfiguration]
-   * is used to configure Dokka behaviour.)
-   */
-  @DokkatooInternalApi
-  class DependencyContainers(
-    private val formatName: String,
-    dokkatooConsumer: NamedDomainObjectProvider<Configuration>,
-    project: Project,
-  ) {
-
-    private val objects: ObjectFactory = project.objects
-
-    private val dependencyContainerNames = DokkatooBasePlugin.DependencyContainerNames(formatName)
-
-    private val dokkatooAttributes: DokkatooConfigurationAttributes = objects.newInstance()
-
-    private fun AttributeContainer.dokkaCategory(category: DokkatooConfigurationAttributes.DokkatooCategoryAttribute) {
-      attribute(DOKKATOO_BASE_ATTRIBUTE, dokkatooAttributes.dokkatooBaseUsage)
-      attribute(DOKKA_FORMAT_ATTRIBUTE, objects.named(formatName))
-      attribute(DOKKATOO_CATEGORY_ATTRIBUTE, category)
-    }
-
-    private fun AttributeContainer.jvmJar() {
-      attribute(USAGE_ATTRIBUTE, objects.named(JAVA_RUNTIME))
-      attribute(CATEGORY_ATTRIBUTE, objects.named(LIBRARY))
-      attribute(BUNDLING_ATTRIBUTE, objects.named(EXTERNAL))
-      attribute(TARGET_JVM_ENVIRONMENT_ATTRIBUTE, objects.named(STANDARD_JVM))
-      attribute(LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(JAR))
-    }
-
-    //<editor-fold desc="Dokka Parameters JSON files">
-    // TODO sharing parameters is required for a 'DokkaCollect' equivalent, but this is not implemented yet
-    /** Fetch Dokka Parameter files from other subprojects */
-    @Suppress("unused", "DEPRECATION")
-    @Deprecated("Not used")
-    val dokkaParametersConsumer: NamedDomainObjectProvider<Configuration> =
-      project.configurations.register(dependencyContainerNames.dokkatooParametersConsumer) {
-        description = "Fetch Dokka Parameters for $formatName from other subprojects"
-        asConsumer()
-        extendsFrom(dokkatooConsumer.get())
-        attributes {
-          dokkaCategory(dokkatooAttributes.dokkaParameters)
-        }
-      }
-
-    /** Provide Dokka Parameter files to other subprojects */
-    @Suppress("unused", "DEPRECATION")
-    @Deprecated("Not used")
-    val dokkaParametersOutgoing: NamedDomainObjectProvider<Configuration> =
-      project.configurations.register(dependencyContainerNames.dokkatooParametersOutgoing) {
-        description = "Provide Dokka Parameters for $formatName to other subprojects"
-        asProvider()
-        attributes {
-          dokkaCategory(dokkatooAttributes.dokkaParameters)
-        }
-      }
-    //</editor-fold>
-
-    //<editor-fold desc="Dokka Module files">
-    /** Fetch Dokka Module files from other subprojects */
-    val dokkaModuleConsumer: NamedDomainObjectProvider<Configuration> =
-      project.configurations.register(dependencyContainerNames.dokkatooModuleFilesConsumer) {
-        description = "Fetch Dokka Module files for $formatName from other subprojects"
-        asConsumer()
-        extendsFrom(dokkatooConsumer.get())
-        attributes {
-          dokkaCategory(dokkatooAttributes.dokkaModuleFiles)
-        }
-      }
-    /** Provide Dokka Module files to other subprojects */
-    val dokkaModuleOutgoing: NamedDomainObjectProvider<Configuration> =
-      project.configurations.register(dependencyContainerNames.dokkatooModuleFilesProvider) {
-        description = "Provide Dokka Module files for $formatName to other subprojects"
-        asProvider()
-        // extend from dokkaConfigurationsConsumer, so Dokka Module Configs propagate api() style
-        extendsFrom(dokkaModuleConsumer.get())
-        attributes {
-          dokkaCategory(dokkatooAttributes.dokkaModuleFiles)
-        }
-      }
-    //</editor-fold>
-
-    //<editor-fold desc="Dokka Generator Plugins">
-    /**
-     * Dokka plugins.
-     *
-     * Users can add plugins to this dependency.
-     *
-     * Should not contain runtime dependencies.
-     */
-    val dokkaPluginsClasspath: NamedDomainObjectProvider<Configuration> =
-      project.configurations.register(dependencyContainerNames.dokkaPluginsClasspath) {
-        description = "Dokka Plugins classpath for $formatName"
-        asConsumer()
-        attributes {
-          jvmJar()
-          dokkaCategory(dokkatooAttributes.dokkaPluginsClasspath)
-        }
-      }
-
-    /**
-     * Dokka Plugins, without transitive dependencies.
-     *
-     * It extends [dokkaPluginsClasspath], so do not add dependencies to this configuration -
-     * the dependencies are computed automatically.
-     */
-    val dokkaPluginsIntransitiveClasspath: NamedDomainObjectProvider<Configuration> =
-      project.configurations.register(dependencyContainerNames.dokkaPluginsIntransitiveClasspath) {
-        description =
-          "Dokka Plugins classpath for $formatName - for internal use. Fetch only the plugins (no transitive dependencies) for use in the Dokka JSON Configuration."
-        asConsumer()
-        extendsFrom(dokkaPluginsClasspath.get())
-        isTransitive = false
-        attributes {
-          jvmJar()
-          dokkaCategory(dokkatooAttributes.dokkaPluginsClasspath)
-        }
-      }
-    //</editor-fold>
-
-    //<editor-fold desc="Dokka Generator Classpath">
-    /**
-     * Runtime classpath used to execute Dokka Worker.
-     *
-     * This configuration is not exposed to other subprojects.
-     *
-     * Extends [dokkaPluginsClasspath].
-     *
-     * @see dev.adamko.dokkatoo.workers.DokkaGeneratorWorker
-     * @see dev.adamko.dokkatoo.tasks.DokkatooGenerateTask
-     */
-    val dokkaGeneratorClasspath: NamedDomainObjectProvider<Configuration> =
-      project.configurations.register(dependencyContainerNames.dokkaGeneratorClasspath) {
-        description =
-          "Dokka Generator runtime classpath for $formatName - will be used in Dokka Worker. Should contain all transitive dependencies, plugins (and their transitive dependencies), so Dokka Worker can run."
-        asConsumer()
-
-        // extend from plugins classpath, so Dokka Worker can run the plugins
-        extendsFrom(dokkaPluginsClasspath.get())
-
-        isTransitive = true
-        attributes {
-          jvmJar()
-          dokkaCategory(dokkatooAttributes.dokkaGeneratorClasspath)
-        }
-      }
-    //</editor-fold>
-  }
-
-  @DokkatooInternalApi
-  class DokkatooTasks(
-    project: Project,
-    private val publication: DokkaPublication,
-    private val dokkatooExtension: DokkatooExtension,
-    private val dependencyContainers: DependencyContainers,
-
-    private val providers: ProviderFactory,
-  ) {
-    private val formatName: String get() = publication.formatName
-
-    private val taskNames = DokkatooBasePlugin.TaskNames(formatName)
-
-    private fun DokkatooGenerateTask.applyFormatSpecificConfiguration() {
-      runtimeClasspath.from(
-        dependencyContainers.dokkaGeneratorClasspath.map { classpath ->
-          classpath.incoming.artifacts.artifactFiles
-        }
-      )
-      generator.apply {
-        publicationEnabled.convention(publication.enabled)
-
-        failOnWarning.convention(publication.failOnWarning)
-        finalizeCoroutines.convention(publication.finalizeCoroutines)
-        includes.from(publication.includes)
-        moduleName.convention(publication.moduleName)
-        moduleVersion.convention(publication.moduleVersion)
-        offlineMode.convention(publication.offlineMode)
-        pluginsConfiguration.addAllLater(providers.provider { publication.pluginsConfiguration })
-        pluginsClasspath.from(
-          dependencyContainers.dokkaPluginsIntransitiveClasspath.map { classpath ->
-            classpath.incoming.artifacts.artifactFiles
-          }
-        )
-        suppressInheritedMembers.convention(publication.suppressInheritedMembers)
-        suppressObviousFunctions.convention(publication.suppressObviousFunctions)
-      }
-    }
-
-    @Suppress("DEPRECATION", "unused")
-    @Deprecated("No longer required")
-    val prepareParameters = project.tasks.register<DokkatooPrepareParametersTask>(
-      taskNames.prepareParameters,
-    ).configuring task@{
-      description =
-        "[DEPRECATED no longer used] Prepares Dokka parameters for generating the $formatName publication"
-
-      dokkaConfigurationJson.convention(
-        dokkatooExtension.dokkatooConfigurationsDirectory.file("$formatName/dokka_parameters.json")
-      )
-    }
-
-    val generatePublication = project.tasks.register<DokkatooGenerateTask>(
-      taskNames.generatePublication,
-      publication.pluginsConfiguration,
-    ).configuring task@{
-      description = "Executes the Dokka Generator, generating the $formatName publication"
-      generationType.set(DokkatooGenerateTask.GenerationType.PUBLICATION)
-
-      outputDirectory.convention(dokkatooExtension.dokkatooPublicationDirectory.dir(formatName))
-
-      generator.apply {
-        // depend on Dokka Module Descriptors from other subprojects
-        dokkaModuleFiles.from(
-          dependencyContainers.dokkaModuleConsumer.map { modules ->
-            modules.incoming
-              .artifactView { componentFilter(LocalProjectOnlyFilter) }
-              .artifacts.artifactFiles
-          }
-        )
-      }
-
-      applyFormatSpecificConfiguration()
-    }
-
-    val generateModule = project.tasks.register<DokkatooGenerateTask>(
-      taskNames.generateModule,
-      publication.pluginsConfiguration,
-    ).configuring task@{
-      description = "Executes the Dokka Generator, generating a $formatName module"
-      generationType.set(DokkatooGenerateTask.GenerationType.MODULE)
-
-      outputDirectory.convention(dokkatooExtension.dokkatooModuleDirectory.dir(formatName))
-
-      applyFormatSpecificConfiguration()
-    }
-
-    val prepareModuleDescriptor = project.tasks.register<DokkatooPrepareModuleDescriptorTask>(
-      taskNames.prepareModuleDescriptor
-    ) task@{
-      description = "Prepares the Dokka Module Descriptor for $formatName"
-      includes.from(publication.includes)
-      dokkaModuleDescriptorJson.convention(
-        dokkatooExtension.dokkatooConfigurationsDirectory.file("$formatName/module_descriptor.json")
-      )
-      moduleDirectory.set(generateModule.flatMap { it.outputDirectory })
-
-//      dokkaSourceSets.addAllLater(providers.provider { dokkatooExtension.dokkatooSourceSets })
-//      dokkaSourceSets.configureEach {
-//        sourceSetScope.convention(this@task.path)
-//      }
-    }
-  }
 }
